requirements.txt
mysql-connector-python
tqdm
linkedin-jobs-scraper
selenium
beautifulsoup4